chore: add to BudgetDetailActivityTabContents
