+++ src/components/learner-credit-management/BudgetDetailActivityTabContents.jsx
@@ -1,19 +1,29 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import NoBudgetActivityCard from './NoBudgetActivityCard';
+import { connect } from 'react-redux';
+import { useParams } from 'react-router-dom';
+
 import LearnerCreditAllocationTable from './LearnerCreditAllocationTable';
+import NoBudgetActivityCard from './NoBudgetActivityCard';
+import { useOfferRedemptions, isUUID } from './data';
 
 const BudgetDetailActivityTabContents = ({
   isTopDownAssignmentRealTimeLcmEnabled,
   hasPendingAssignments,
   hasCompletedTransactions,
-  isLoadingOfferRedemptions,
-  fetchOfferRedemptions,
-  offerRedemptions,
   enterpriseUUID,
   enterpriseSlug,
   enableLearnerPortal,
 }) => {
+  const { budgetId } = useParams();
+  const enterpriseOfferId = isUUID(budgetId) ? null : budgetId;
+  const subsidyAccessPolicyId = isUUID(budgetId) ? budgetId : null;
+  const {
+    isLoading: isLoadingOfferRedemptions,
+    offerRedemptions,
+    fetchOfferRedemptions,
+  } = useOfferRedemptions(enterpriseUUID, enterpriseOfferId, subsidyAccessPolicyId);
+
   if (!isTopDownAssignmentRealTimeLcmEnabled) {
     return (
       <LearnerCreditAllocationTable
@@ -67,19 +77,24 @@ const BudgetDetailActivityTabContents = ({
   return null;
 };
 
+const mapStateToProps = state => ({
+  enterpriseUUID: state.portalConfiguration.enterpriseId,
+  enterpriseSlug: state.portalConfiguration.enterpriseSlug,
+  enableLearnerPortal: state.portalConfiguration.enableLearnerPortal,
+});
+
 BudgetDetailActivityTabContents.propTypes = {
+  enterpriseUUID: PropTypes.string.isRequired,
+  enterpriseSlug: PropTypes.string.isRequired,
+  enableLearnerPortal: PropTypes.bool.isRequired,
   isTopDownAssignmentRealTimeLcmEnabled: PropTypes.bool.isRequired,
   hasPendingAssignments: PropTypes.bool.isRequired,
   hasCompletedTransactions: PropTypes.bool.isRequired,
-  isLoadingOfferRedemptions: PropTypes.bool.isRequired,
   offerRedemptions: PropTypes.shape({
     itemCount: PropTypes.number,
     pageCount: PropTypes.number,
     results: PropTypes.arrayOf(PropTypes.object),
   }).isRequired,
-  fetchOfferRedemptions: PropTypes.func.isRequired,
-  enterpriseUUID: PropTypes.string.isRequired,
-  enterpriseSlug: PropTypes.string.isRequired,
-  enableLearnerPortal: PropTypes.bool.isRequired,
 };
-export default BudgetDetailActivityTabContents;
+
+export default connect(mapStateToProps)(BudgetDetailActivityTabContents);
